Changed the way dependencies are detected
If no ViennaHRLE_DIR (or VTK_DIR) is provided, CMake now automatically
tries to detect the ViennaHRLE (and VTK if enabled) installations.
If the provided directory contains the installation, it is used.
Otherwise cmake throws an error. If no directory is provided, cmake
first tries to find the installation in the system paths. If no
existing installation is found, the dependency will be build from source
using the existing externalproject build configurations.